Add helper methods for instance configuration class
