+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/InstanceConfigHelper.java
@@ -1,0 +1,6 @@
+package com.appsmith.server.helpers;
+
+import com.appsmith.server.helpers.ce.InstanceConfigHelperCE;
+
+public interface InstanceConfigHelper extends InstanceConfigHelperCE {
+}
