+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/ce/InstanceConfigHelperCE.java
@@ -1,0 +1,17 @@
+package com.appsmith.server.helpers.ce;
+
+import com.appsmith.server.domains.Config;
+import reactor.core.publisher.Mono;
+
+public interface InstanceConfigHelperCE {
+
+    Mono<? extends Config> registerInstance();
+
+    Mono<Config> checkInstanceSchemaVersion();
+
+    void printReady();
+
+    Mono<Void> performRtsHealthCheck();
+
+    boolean getIsRtsAccessible();
+}
